feat: split produce batch for MESSAGE_TOO_LARGE
Aggregating the records for produce requests is approximate, and as such
sometimes may result in a batch larger than the configured maximum - if
the size of the batch exceeds the broker's maximum allowed message size
(message.max.bytes) a MESSAGE_TOO_LARGE error is returned.

At this point and prior to this commit, the whole batch was dropped and
the caller has to retry, hoping the retries do not batch into a
too-large batch once again. After this commit, exactly one attempt is
made to automatically/internally split the batch in half and attempt to
concurrently produce each half transparently to the caller. A message is
logged at WARN to help the user identify and tune this config to avoid
this.

The two downsides of this are:

    * Unpredictable latency if a batch needs splitting and re-submitting
    * A partial write may occur (writing one half succeeds, other fails)

These downsides are offset by the fact the MESSAGE_TOO_LARGE should be a
rare event, making the latency hit equally rare. Succeeding in producing
one batch but not the other is even rarer still, but does result in an
error being returned to the client even though have the records were
successfully produced. The caller had no control over which writes were
in which aggregated batches anyway, so has no expectation of two writes
succeeding/failing together (atomicity).