refactor: let MockClient return response sequences
This change affects test code only, changing the MockClient to return a
sequence of configured errors, rather than the same error all the time.

As an example, this can be used to construct a test that observes an
error, followed by a different error, followed by a success - the
previous mock was restricted to always returning the same response.