fix: use compact arrays where needed
Compact array representations were introduced at the same time as
compact strings and tagged fields. For arrays with versioned structures
we already use these. For primitive arrays however we did not use them
yet. The only message where this applies is CreateTopics and the broker
didn't seem to care that we've messed up the serialization of this one
field.

Compact arrays might be helpful for other message types as well, so this
is both a fix and some feature work.