feat: support pre-warming PartitionClient
Adds a refresh_leader() method to force the PartitionClient through the
discovery process.

This will ensure a cached connection to the leader is (likely) already
established by the time the first produce() call is issued, preventing
the first produce() calls (multiple of which may be queued up on the
discovery mutex) from experiencing excessively high latency.