perf: async batch flushing & lock contention
Rewrites the management of the aggregation buffer to enable asynchronous
flushing of batched writes, while continuing to build up the next buffer
of writes in parallel.

A behavioural change of note: previously there was (synchronous)
back-pressure applied when a batch was being flushed, and there could
only ever be at most 1 batch of writes outstanding. After this commit
this back-pressure is removed as a side effect of async flushing, and
there is no upper bound on the number of in-flight batch writes
enforced. Back-pressure can easily be placed back into rskafka if it is
deemed the responsibility of rskafka, rather than the caller, to limit
the number of outstanding produce() calls.

This commit also eliminates lock contention between the callers
attempting to add to the batch, and those that are waiting for their
write result, and also minimises the spurious task wakeups caused by all
callers waiting on the linger timer, before (likely) moving on to wait
on the async mutex, before finally being able to read their write result
over the broadcast channel (also a mutex). Now one caller performs a
linger dance, while all the others wait exclusively on their broadcast
handle for the results, avoiding having to wait on the aggregation
mutex.