perf: use RwLock for BroadcastOnce
Prior to this commit, all the readers (holders of BroadcastOnceReceiver)
would contend with each other to read the value that has been broadcast.
The BroadcastOnce type restricts usage to one writer, with many readers,
making the case of >1 reader contending to read the result common by
design.

A RwLock eliminates the reader contention, preventing the readers from
queuing up to read their value - it is now a pair of cheap atomic
inc/sub to read the value, in parallel, for each reader.